PASS logout tests
